requirements.txt
strsimpy
confusable_homoglyphs
homoglyphs
colorama
requests
numpy
pytest
pytest-cov
codecov
coverage
black
flake8
beautifulsoup4
discord
python-dotenv
dnspython==1.16.0